+++ NEWS.md
@@ -5,7 +5,15 @@ SHELF v1.1.0 (2016-01-29)
 
 * New functions added for eliciting beliefs about uncertain population distributions:  cdffeedback, cdfplot, fitprecision, pdfplots
 
-SHELF v1.1.0.9001 (2016-04-21) 
+SHELF v1.1.0.9001 (2016-07-06) 
 ==============================
 
 * New function shinyRoulette for using the roulette elicitation method within shiny
+
+* New functions fitDirchlet and feedbackDirichlet for eliciting Dirichlet distributions
+
+* New functions copulaSample and elicitQuadProb for eliciting dependent distributions using multivariate normal copulas
+
+* Change to expert.names from numbers to letters in fitdist
+
+* Vignettes added: overview of SHELF, eliciting a Dirichlet distribution, eliciting a bivariate distribution with a bivariate normal copula